+++ src/main/java/com/example/pamphboard/util/Pagination.java
@@ -1,4 +1,4 @@
-package com.example.pamphboard.dto;
+package com.example.pamphboard.util;
 
 public class Pagination {
